+++ atomate/vasp/tests/test_drones.py
@@ -12,6 +12,8 @@ from pymatgen.io.vasp import Outcar, Oszicar
 
 from atomate.vasp.drones import VaspDrone
 
+import numpy as np
+
 
 module_dir = os.path.join(os.path.dirname(os.path.abspath(__file__)))
 
@@ -68,6 +70,37 @@ class VaspToDbTaskDroneTest(unittest.TestCase):
 
     def test_bandstructure(self):
         drone = VaspDrone()
+
+        doc = drone.assimilate(self.relax2)
+        self.assertEqual(doc["composition_reduced"], {'Si': 1.0})
+        self.assertEqual(doc["formula_pretty"], 'Si')
+        self.assertEqual(doc["formula_anonymous"], 'A')
+        for d in [doc["calcs_reversed"][0]["output"], doc["output"]]:
+            self.assertAlmostEqual(d["vbm"],5.6147)
+            self.assertAlmostEqual(d["cbm"],6.2652)
+            self.assertAlmostEqual(d["bandgap"], 0.6505)
+            self.assertFalse(d["is_gap_direct"])
+            self.assertFalse(d["is_metal"])
+            self.assertNotIn("transition",d)
+            self.assertAlmostEqual(d["direct_gap"],2.5561)
+            self.assertNotIn("bandstructure",doc["calcs_reversed"][0])
+
+
+        doc = drone.assimilate(self.Si_static)
+        self.assertEqual(doc["composition_reduced"], {'Si': 1.0})
+        self.assertEqual(doc["formula_pretty"], 'Si')
+        self.assertEqual(doc["formula_anonymous"], 'A')
+        for d in [doc["calcs_reversed"][0]["output"], doc["output"]]:
+            self.assertAlmostEqual(d["vbm"],5.6138)
+            self.assertAlmostEqual(d["cbm"],6.2644)
+            self.assertAlmostEqual(d["bandgap"],  0.6506)
+            self.assertFalse(d["is_gap_direct"])
+            self.assertFalse(d["is_metal"])
+            self.assertNotIn("transition",d)
+            self.assertAlmostEqual(d["direct_gap"],2.5563)
+            self.assertIn("bandstructure", doc["calcs_reversed"][0])
+
+
         doc = drone.assimilate(self.Al)
         self.assertEqual(doc["composition_reduced"], {'Al': 1.0})
         self.assertEqual(doc["formula_pretty"], 'Al')
@@ -78,6 +111,7 @@ class VaspToDbTaskDroneTest(unittest.TestCase):
             self.assertEqual(d["bandgap"], 0.0)
             self.assertFalse(d["is_gap_direct"])
             self.assertTrue(d["is_metal"])
+            self.assertEqual(doc["calcs_reversed"][0]["bandstructure"]["@class"],"BandStructureSymmLine")
 
     def test_detect_output_file_paths(self):
         drone = VaspDrone()
@@ -97,6 +131,20 @@ class VaspToDbTaskDroneTest(unittest.TestCase):
         self.assertDictEqual({'chgcar': 'CHGCAR.relax1.gz', 'procar': 'PROCAR.relax1.gz',
                               'wavecar': 'WAVECAR.relax1.gz'},
                              doc['calcs_reversed'][1]['output_file_paths'])
+
+    def test_parse_locpot(self):
+        drone = VaspDrone(parse_locpot=True)
+        doc = drone.assimilate(self.Si_static)
+
+        self.assertTrue(drone.parse_locpot)
+        self.assertTrue('locpot' in doc['calcs_reversed'][0]['output'])
+        self.assertTrue(0 in doc['calcs_reversed'][0]['output']['locpot'])
+        self.assertTrue(1 in doc['calcs_reversed'][0]['output']['locpot'])
+        self.assertTrue(2 in doc['calcs_reversed'][0]['output']['locpot'])
+
+        self.assertAlmostEqual(np.sum(doc['calcs_reversed'][0]['output']['locpot'][0]),0)
+        self.assertAlmostEqual(np.sum(doc['calcs_reversed'][0]['output']['locpot'][1]),0)
+        self.assertAlmostEqual(np.sum(doc['calcs_reversed'][0]['output']['locpot'][2]),0)
 
 if __name__ == "__main__":
     unittest.main()