+++ docs_rst/changelog.rst
@@ -2,6 +2,47 @@
 atomate Changelog
 =================
 
+**v0.6.9**
+
+* update requirements to include boltons
+* bugfix for atwf (S. Dwaraknath)
+
+**v0.6.8**
+
+* New SCAN functional workflow (S. Dwaraknath)
+* remove dependence on pymatgen-db (S. Dwaraknath)
+* more bandgap properties parsed by drone (transition and is-direct) (S. Dwaraknath)
+* option to clean up large output files like WAVECAR (S. Dwaraknath)
+* option to recursively copy file tree in CopyFilesFromCalcLoc (A. Faghaninia)
+* bugfix: apply vasp_input_set_params when StaticFW have parents (specter119)
+* misc bugfixes (S. Dwaraknath, A. Jain)
+
+**v0.6.7**
+
+* New ferroelectrics workflow! (T. Smidt)
+* option to parse LOCPOT in VaspDrone (S. Dwaraknath)
+* rename set_fworker -> set_execution_option
+* more options for BoltztrapFW (A. Faghaninia)
+* misc. bugfixes (D. Broberg, K. Mathew, P. Huck)
+
+**v0.6.6**
+
+* powerup to preserve the same FWorker for all jobs in workflow (S. Dwaraknath)
+* DriftErrorHandler in VASP custodian jobs (S. Dwaraknath)
+* some FireTasks in anticipation of ferroelectrics workflow (T. Schmidt, A. Jain)
+
+**v0.6.5**
+
+* fix delta_volume_percent, set as new key and update FixTasksBuilder (B. Bocklund, A. Jain)
+* drone schema version reflects atomate version (M. Horton)
+* unit test fix (J. Montoya)
+
+**v0.6.4**
+
+* add config option for half_kpts_first and max force (A. Jain, S. Dwaraknath)
+* better logic for band structure parsing (S. Dwaraknath)
+* misc bugfix (P. Huck)
+
 **v0.6.3**
 
 * fix Gibbs wf db insertion (A. Dunn, K. Mathew)